Added board file to bin folder and call files from main.rb
